+++ src/pages/Terms/ContentPaper.tsx
