+++ src/pages/Terms/NestedList.tsx
@@ -1,0 +1,8 @@
+import styled from 'styled-components/macro';
+import List from '@material-ui/core/List';
+
+const NestedList = styled(List)`
+  margin-left: 17px;
+`;
+
+export default NestedList;
