+++ src/pages/Terms/Sentence.tsx
@@ -1,0 +1,7 @@
+import styled from 'styled-components/macro';
+
+const Sentence = styled.div`
+  padding: 17px;
+`;
+
+export default Sentence;
